@@ -16,21 +16,28 @@ void test() {
     int sock = ::socket(AF_INET, SOCK_STREAM, 0);
     ::fcntl(sock, F_SETFL, O_NONBLOCK);
 
-    sockaddr_in addr;
+    ::sockaddr_in addr;
     ::memset(&addr, 0, sizeof(addr));
     addr.sin_family = AF_INET;
     addr.sin_port = htons(80);
-    inet_pton(AF_INET, "14.215.177.39", &addr.sin_addr.s_addr);
+    ::inet_pton(AF_INET, "14.215.177.39", &addr.sin_addr.s_addr);
 
-    if(!connect(sock, (const sockaddr*)&addr, sizeof(addr))) {
+    if(!::connect(sock, (const sockaddr*)&addr, sizeof(sockaddr))) {
     } else if(errno == EINPROGRESS) {
         QFF_LOG_INFO(QFF_LOG_ROOT) << "add event errno=" << errno << " " << strerror(errno);
-        IOManager::GetThis()->add_event(sock, E_T_READ, [](){
+        IOManager::GetThis()->add_event(sock, IOManager::READ, [=](){
             QFF_LOG_INFO(QFF_LOG_ROOT) << "read callback";
+            //char buf[1024];
+            //::recv(sock, buf, 1024, 0);
+           // close(sock);
+            //QFF_LOG_INFO(QFF_LOG_ROOT) << buf;
         });
-        IOManager::GetThis()->add_event(sock, E_T_WRITE, [=](){
+        IOManager::GetThis()->add_event(sock, IOManager::WRITE, [=](){
             QFF_LOG_INFO(QFF_LOG_ROOT) << "write callback";
-            IOManager::GetThis()->cancel_event(sock, E_T_READ);
+            //char buf[] = "GET / HTTP/1.1\r\n"
+            //             "Host:www.baidu.com\r\n\r\n";
+            //::send(sock, buf, sizeof(buf), 0);
+            IOManager::GetThis()->cancel_event(sock, IOManager::READ);
             close(sock);
         });
     } else {
@@ -41,9 +48,13 @@ void test() {
 int main() {
     LoggerMgr::New();
 
-    IOManager::ptr iom = std::make_shared<IOManager>(1, "main");
+    IOManager::ptr iom = std::make_shared<IOManager>(2, "main");
     iom->schedule(test);
-    iom->start();
+    iom->schedule(test);
+    iom->schedule(test);
+    iom->schedule(test);
+    iom->schedule(test);
+    iom->schedule(test);
     QFF_LOG_DEBUG(QFF_LOG_ROOT) << "789";
     iom->stop();
     LoggerMgr::Delete();
